+++ release/i18next-1.6.3.js
@@ -1,20 +1,8 @@
-// i18next, v1.6.3_pre
+// i18next, v1.6.3
 // Copyright (c)2013 Jan Mühlemann (jamuhl).
 // Distributed under MIT license
 // http://i18next.com
-(function (root, factory) {
-    if (typeof exports === 'object') {
-
-      var jquery = require('jquery');
-
-      module.exports = factory(jquery);
-
-    } else if (typeof define === 'function' && define.amd) {
-
-      define(['jquery'], factory);
-
-    } 
-}(this, function ($) {
+(function() {
 
     // add indexOf to non ECMA-262 standard compliant browsers
     if (!Array.prototype.indexOf) {
@@ -81,12 +69,27 @@
         };
     }
 
-    var i18n = {}
-        , resStore = {}
-        , currentLng
-        , replacementCounter = 0
-        , languages = [];
+    var root = this
+      , $ = root.jQuery || root.Zepto
+      , i18n = {}
+      , resStore = {}
+      , currentLng
+      , replacementCounter = 0
+      , languages = [];
 
+
+    // Export the i18next object for **CommonJS**. 
+    // If we're not in CommonJS, add `i18n` to the
+    // global object or to jquery.
+    if (typeof module !== 'undefined' && module.exports) {
+        module.exports = i18n;
+    } else {
+        if ($) {
+            $.i18n = $.i18n || i18n;
+        }
+        
+        root.i18n = root.i18n || i18n;
+    }
     // defaults
     var o = {
         lng: undefined,
@@ -2607,9 +2610,4 @@
     i18n.addPostProcessor = addPostProcessor;
     i18n.options = o;
 
-    $.i18n = i18n;
-    $.t = i18n.t;
-        
-    return i18n;
-
-}));
+})();