+++ release/i18next.amd-1.6.3.js
@@ -1,8 +1,18 @@
-// i18next, v1.6.3_pre
+// i18next, v1.6.3
 // Copyright (c)2013 Jan Mühlemann (jamuhl).
 // Distributed under MIT license
 // http://i18next.com
-(function() {
+(function (root, factory) {
+    if (typeof exports === 'object') {
+
+        module.exports = factory();
+
+    } else if (typeof define === 'function' && define.amd) {
+
+        define([], factory);
+
+    } 
+}(this, function () {
 
     // add indexOf to non ECMA-262 standard compliant browsers
     if (!Array.prototype.indexOf) {
@@ -69,27 +79,13 @@
         };
     }
 
-    var root = this
-      , $ = root.jQuery || root.Zepto
-      , i18n = {}
-      , resStore = {}
-      , currentLng
-      , replacementCounter = 0
-      , languages = [];
+    var $ = undefined
+        , i18n = {}
+        , resStore = {}
+        , currentLng
+        , replacementCounter = 0
+        , languages = [];
 
-
-    // Export the i18next object for **CommonJS**. 
-    // If we're not in CommonJS, add `i18n` to the
-    // global object or to jquery.
-    if (typeof module !== 'undefined' && module.exports) {
-        module.exports = i18n;
-    } else {
-        if ($) {
-            $.i18n = $.i18n || i18n;
-        }
-        
-        root.i18n = root.i18n || i18n;
-    }
     // defaults
     var o = {
         lng: undefined,
@@ -2609,5 +2605,7 @@
     i18n.lng = lng;
     i18n.addPostProcessor = addPostProcessor;
     i18n.options = o;
+        
+    return i18n; 
 
-})();
+}));