+++ spec/init/init.namespace.spec.js
@@ -130,7 +130,7 @@ describe('with namespace', function() {
           i18n.init(i18n.functions.extend(opts, { 
             fallbackNS: ['ns.fallback1', 'ns.fallback2'], 
             resStore: resStore,
-            sendMissing: true,
+            sendMissing: true, /* must be changed to saveMissing */
             ns: { namespaces: ['ns.common', 'ns.special', 'ns.fallback'], defaultNs: 'ns.special'} } ),
             function(t) { 
               t('ns.common:notExisting');