+++ spec/server.functions.spec.js
@@ -16,6 +16,8 @@ describe('i18next.functions', function() {
     opts = {
       lng: 'en-US',
       fallbackLng: 'dev',
+      fallbackNS: [],
+      fallbackToDefaultNS: false,
       load: 'all',
       preload: [],
       supportedLngs: [],