+++ spec/server.init.spec.js
@@ -2,6 +2,7 @@
 // HINT
 //
 // you need to replace '_fetchOne' with 'fetchOne' to use this on server
+// fix line 351 'sendMissing' -> 'saveMissing'
 //
 
 
@@ -17,6 +18,8 @@ describe('i18next.init', function() {
     opts = {
       lng: 'en-US',
       fallbackLng: 'dev',
+      fallbackNS: [],
+      fallbackToDefaultNS: false,
       load: 'all',
       preload: [],
       supportedLngs: [],